requirements.txt
# requirements.txt

opencv-python
numpy
Pillow
ttkthemes